Refactor functional options to avoid using internal consts; this also prevents accidentally passing invalid numerical values
