+++ autolayout_options.go
@@ -40,69 +40,3 @@ var defaultOptions = options{
 }
 
 type Option func(*options)
-
-func WithCycleBreaking(alg cbreaking.Alg) Option {
-	return func(o *options) {
-		o.p1 = alg
-	}
-}
-
-func WithLayering(alg layering.Alg) Option {
-	return func(o *options) {
-		o.p2 = alg
-	}
-}
-
-func WithOrdering(alg ordering.Alg) Option {
-	return func(o *options) {
-		o.p3 = alg
-	}
-}
-
-func WithPositioning(alg positioning.Alg) Option {
-	return func(o *options) {
-		o.p4 = alg
-	}
-}
-
-func WithEdgeRouting(alg routing.Alg) Option {
-	return func(o *options) {
-		o.p5 = alg
-	}
-}
-
-func WithNetworkSimplexThoroughness(thoroughness uint) Option {
-	return func(o *options) {
-		o.params.NetworkSimplexThoroughness = thoroughness
-	}
-}
-
-func WithNetworkSimplexBalance(balance graph.OptionNsBalance) Option {
-	return func(o *options) {
-		o.params.NetworkSimplexBalance = balance
-	}
-}
-
-func WithLayerSpacing(spacing float64) Option {
-	return func(o *options) {
-		o.params.LayerSpacing = spacing
-	}
-}
-
-func WithNodeSpacing(spacing float64) Option {
-	return func(o *options) {
-		o.params.NodeSpacing = spacing
-	}
-}
-
-func WithBrandesKoepfLayout(i int) Option {
-	return func(o *options) {
-		o.params.BrandesKoepfLayout = i
-	}
-}
-
-func WithMonitor(monitor imonitor.Monitor) Option {
-	return func(o *options) {
-		o.monitor = monitor
-	}
-}
