+++ autolayout_options_test.go
@@ -5,7 +5,6 @@ import (
 
 	"github.com/nulab/autog/internal/graph"
 	"github.com/nulab/autog/internal/phase1"
-	"github.com/nulab/autog/internal/phase2"
 	"github.com/nulab/autog/internal/phase3"
 	"github.com/nulab/autog/internal/phase4"
 	"github.com/nulab/autog/internal/phase5"
@@ -15,11 +14,10 @@ import (
 func TestOptions(t *testing.T) {
 	// set some random options that are different from the defaults
 	opts := testOptions(
-		WithCycleBreaking(phase1.DepthFirst),
-		WithLayering(phase2.Alg(12)),
-		WithOrdering(phase3.NoOrdering),
-		WithPositioning(phase4.VerticalAlign),
-		WithEdgeRouting(phase5.Straight),
+		WithCycleBreakingDFS(),
+		WithOrderingNoop(),
+		WithPositioningVAlign(),
+		WithEdgeRoutingStraight(),
 		WithNetworkSimplexThoroughness(30),
 		WithNetworkSimplexBalance(graph.OptionNsBalanceH),
 		WithLayerSpacing(75.5),
@@ -28,7 +26,6 @@ func TestOptions(t *testing.T) {
 	)
 
 	assert.Equal(t, phase1.DepthFirst, opts.p1)
-	assert.Equal(t, phase2.Alg(12), opts.p2)
 	assert.Equal(t, phase3.NoOrdering, opts.p3)
 	assert.Equal(t, phase4.VerticalAlign, opts.p4)
 	assert.Equal(t, phase5.Straight, opts.p5)
